feat(pokemon): implement pokemon list loading effect
This commit introduces an NgRx effect to handle the loading of pokemons.

The PokemonEffects class is added, which includes the loadPokemons$ effect. This effect listens for the PokemonActions.loadPokemons action, calls the pokemonService.getPokemonList to fetch pokemon data, and then dispatches either PokemonActions.loadPokemonsSuccess with the fetched data or PokemonActions.loadPokemonsFailure in case of an error.